+++ src/components/Link/styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const Link = styled.a`
+  color: palevioletred;
+
+  &:visited {
+    color: palegoldenrod;
+  }
+`
+
+export { Link }