+++ src/utils/firebase/index.ts
@@ -1,6 +1,13 @@
 import * as firebase from 'firebase'
 const uuidv4 = require('uuid/v4')
 
+const signOut = () => {
+  firebase
+    .auth()
+    .signOut()
+    .then()
+}
+
 const sendComment = ({
   content,
   author,
@@ -42,4 +49,39 @@ const readComments = () => {
   })
 }
 
-export { sendComment, readComments }
+const deleteCommentById = (id: string) => {
+  firebase
+    .database()
+    .ref(`comments/${id}`)
+    .remove()
+    .then(() => {})
+}
+
+/**
+ * **updateComment** is a method used to update the value
+ * of a comment. Keep in mind that the `id` value we are
+ * using is not the `comment.UUID` but the object key
+ * within the FB database.
+ *
+ * Since FB stores everything as an object, the way we
+ * identify the comment is through the comment's object key
+ * which is a unique ID. Our front end never uses that key
+ * except for when we interact with firebase through
+ * a couple methods
+ *
+ * @TODO add link to github for more information
+ *
+ * @param id obj key in FB DB, not the comment.UUID
+ * @param content new value of comment
+ */
+const updateComment = (id: string, content: string) => {
+  firebase
+    .database()
+    .ref()
+    .update({
+      [`comments/${id}/content`]: content,
+    })
+    .then(() => {})
+}
+
+export { sendComment, readComments, deleteCommentById, updateComment, signOut }