+++ CHANGELOG.md
@@ -6,13 +6,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## [0.0.7] - 2019-06-30
+
+### Added
+
+- Icons package
+- comment delete feature
+- "update comment" feature and associated components
+- Link component
+- settings menu
+
+### Changes
+
+- significant styling updates to how comments are rendered, now handles overflow correctly, renders an icon area
+
 ## [0.0.2] - 2019-06-24
 
 ### Added
 
 - added Loader component
 - updated sign in / get session process to hopefully render a loading state when fetching the user
-- added more files to npmignore
+- added more files to `npmignore`
 - updated Feedback to subscribe to comments and update state when comments length from firebase does not equal comments in local state
 - updated sign in logic to prevent infinite renders
 - replaced "refresh" button with "clear" button